+++ src/main/java/net/zhaiji/chestcavitybeyond/event/CommonEventManager.java
@@ -25,6 +25,7 @@ public class CommonEventManager {
     public static void gameBusListener(IEventBus gameBus) {
         gameBus.addListener(CommonEventHandler::handlerEntityJoinLevelEvent);
         gameBus.addListener(CommonEventHandler::handlerPlayerEvent$PlayerRespawnEvent);
+        gameBus.addListener(CommonEventHandler::handlerPlayerInteractEvent$EntityInteractSpecific);
         gameBus.addListener(CommonEventHandler::handlerPlayerInteractEvent$EntityInteract);
         gameBus.addListener(CommonEventHandler::handlerMobEffectEvent$Applicable);
         gameBus.addListener(CommonEventHandler::handlerRegisterBrewingRecipesEvent);